+++ mipac/manager/admins/invite.py
@@ -1,0 +1,19 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
+from mipac.abstract.manager import AbstractManager
+from mipac.http import HTTPClient
+from mipac.actions.admins.invite import AdminInviteActions
+
+if TYPE_CHECKING:
+    from mipac.manager.client import ClientManager
+
+
+class AdminInviteManager(AbstractManager):
+    def __init__(self, *, session: HTTPClient, client: ClientManager):
+        self.__session: HTTPClient = session
+        self.__client: ClientManager = client
+
+    @property
+    def action(self) -> AdminInviteActions:
+        return AdminInviteActions(session=self.__session, client=self.__client)